+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/CreateGroupCallOptions.cs
@@ -18,7 +18,6 @@ namespace Azure.Communication.CallAutomation
         {
             Targets = targets;
             CallbackUri = callbackUri;
-            CustomContext = new CustomContext(sipHeaders: new Dictionary<string, string>(), voipHeaders: new Dictionary<string, string>());
         }
 
         /// <summary>
@@ -48,18 +47,8 @@ namespace Azure.Communication.CallAutomation
         public string OperationContext { get; set; }
 
         /// <summary>
-        /// Media Streaming Configuration.
+        /// The endpoint URL of the Azure Cognitive Services resource attached.
         /// </summary>
-        public MediaStreamingOptions MediaStreamingOptions { get; set; }
-
-        /// <summary>
-        /// The endpoint URL of the Azure Cognitive Services resource attached
-        /// </summary>
-        public Uri AzureCognitiveServicesEndpointUri { get; set; }
-
-        /// <summary>
-        /// The Custom Context which contains SIP and voip headers.
-        /// </summary>
-        public CustomContext CustomContext { get; }
+        public Uri CognitiveServicesEndpoint { get; set; }
     }
 }