+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/Events/CallAutomationEventParser.cs
@@ -130,28 +130,10 @@ namespace Azure.Communication.CallAutomation
                     return ContinuousDtmfRecognitionToneFailed.Deserialize(eventData);
                 case nameof(ContinuousDtmfRecognitionStopped):
                     return ContinuousDtmfRecognitionStopped.Deserialize(eventData);
-                case nameof(SendDtmfCompleted):
-                    return SendDtmfCompleted.Deserialize(eventData);
-                case nameof(SendDtmfFailed):
-                    return SendDtmfFailed.Deserialize(eventData);
-                #region Dialog
-                case nameof(DialogCompleted):
-                    return DialogCompleted.Deserialize(eventData);
-                case nameof(DialogFailed):
-                    return DialogFailed.Deserialize(eventData);
-                case nameof(DialogConsent):
-                    return DialogConsent.Deserialize(eventData);
-                case nameof(DialogStarted):
-                    return DialogStarted.Deserialize(eventData);
-                case nameof(DialogHangup):
-                    return DialogHangup.Deserialize(eventData);
-                case nameof(DialogTransfer):
-                    return DialogTransfer.Deserialize(eventData);
-                case nameof(DialogSensitivityUpdate):
-                    return DialogSensitivityUpdate.Deserialize(eventData);
-                case nameof(DialogLanguageChange):
-                    return DialogLanguageChange.Deserialize(eventData);
-                #endregion
+                case nameof(SendDtmfTonesCompleted):
+                    return SendDtmfTonesCompleted.Deserialize(eventData);
+                case nameof(SendDtmfTonesFailed):
+                    return SendDtmfTonesFailed.Deserialize(eventData);
                 default:
                     return null;
             }