+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/RecognizeResultType.cs
@@ -9,7 +9,7 @@ namespace Azure.Communication.CallAutomation
     /// <summary>
     /// Determines the sub-type of the recognize result.
     /// </summary>
-    public readonly partial struct RecognizeResultType: IEquatable<RecognizeResultType>
+    public readonly partial struct RecognizeResultType : IEquatable<RecognizeResultType>
     {
         private readonly string _value;
 
@@ -21,12 +21,9 @@ namespace Azure.Communication.CallAutomation
         }
 
         private const string DtmfResultValue = "DtmfResultValue";
-        private const string CollectTonesResultValue = "CollectTonesResultValue";
         private const string ChoiceResultValue = "ChoiceResultValue";
         private const string SpeechResultValue = "SpeechResultValue";
 
-        /// <summary> CollectTonesResult. </summary>
-        public static RecognizeResultType CollectTonesResult { get; } = new RecognizeResultType(CollectTonesResultValue);
         /// <summary> DtmfResult. </summary>
         public static RecognizeResultType DtmfResult { get; } = new RecognizeResultType(DtmfResultValue);
         /// <summary> ChoiceResult. </summary>