+++ sdk/communication/Azure.Communication.CallAutomation/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.1.0-beta.1 (Unreleased)
+## 1.1.0 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,33 @@
 ### Bugs Fixed
 
 ### Other Changes
+- Dependency versions updated.
+
+## 1.1.0-beta.1 (2023-08-17)
+
+### Features Added
+- Play and recognize supports TTS and SSML source prompts.
+- Recognize supports choices and freeform speech.
+- Start/Stop continuous DTMF recognition by subscribing/unsubscribing to tones.
+- Send DTMF tones to a participant in the call.
+- Mute participants in the call.
+
+## 1.0.0 (2023-06-14)
+
+### Features Added 
+- Outbound calls can now be created without providing a User Identifier. This value can be specified in the CallAutomationClientOption if desired.
+- AnswerCall now accepts OperationContext.
+- Calls can be answered by a specific communication identifier user.
+- RemoveParticipant now sends success and failure events with the request.
+- ParticipantsUpdated event now includes a sequence number to distinguish the ordering of events.
+- CallConnectionProperties now includes CorrelationId.
+- StartRecording now accepts ChannelAffinity.
+- Added EventProcessor, an easy and powerful way to handle Call Automation events. See README for details.
+
+### Breaking Changes 
+- AddParticipant and RemoveParticipant now only accept one participant at a time. 
+- CallSource has been flattened out. 
+- CallInvite model replaces previous models for handling outbound calls.
 
 ## 1.0.0 (2023-06-14)